+++ src/Sender.cpp
@@ -89,6 +89,27 @@ constexpr uint32_t kInterSlotTimerAdjust = 2;
 constexpr uint32_t kInterSlotTimerAdjust = 0;
 #endif  // Which chip?
 
+// Empirically observed MBB timer adjustment constants. The timer adjust values
+// are subtracted from the requested value to get the actual value.
+#if defined(__MK20DX128__) || defined(__MK20DX256__)
+// Min = 24
+constexpr uint32_t kMBBTimerAdjust = 29;
+#elif defined(__MKL26Z64__)
+// Min = 119
+constexpr uint32_t kMBBTimerAdjust = 133;  // At 4us
+#elif defined(__MK64FX512__)
+// Min = 19
+constexpr uint32_t kMBBTimerAdjust = 22;
+#elif defined(__MK66FX1M0__)
+// Min = 11
+constexpr uint32_t kMBBTimerAdjust = 13;
+#elif defined(__IMXRT1062__) || defined(__IMXRT1052__)
+// Min = 2
+constexpr uint32_t kMBBTimerAdjust = 4;
+#else
+constexpr uint32_t kMBBTimerAdjust = 0;
+#endif  // Which chip?
+
 // TX ISR routines
 #if defined(HAS_KINETISK_UART0) || defined(HAS_KINETISL_UART0)
 void uart0_tx_isr();
@@ -178,6 +199,8 @@ Sender::Sender(HardwareSerial &uart)
       adjustedInterSlotTime_(0),
       activePacketSize_(kMaxDMXPacketSize),
       inactivePacketSize_(kMaxDMXPacketSize),
+      mbbTime_(0),
+      adjustedMBBTime_(0),
       refreshRate_(std::numeric_limits<float>::infinity()),
       breakToBreakTime_(0),
       breakStartTime_(0),
@@ -575,6 +598,19 @@ bool Sender::fill(int startChannel, int len, uint8_t value) {
     std::fill_n(&activeBuf_[startChannel], len, value);
   //}
   return true;
+}
+
+void Sender::setMBBTime(uint32_t t) {
+  mbbTime_ = t;
+  if (t <= kMBBTimerAdjust) {
+    adjustedMBBTime_ = 0;
+  } else {
+    adjustedMBBTime_ = t - kMBBTimerAdjust;
+  }
+}
+
+uint32_t Sender::mbbTime() const {
+  return mbbTime_;
 }
 
 bool Sender::setRefreshRate(float rate) {